Improve GC and try_from PyObject tests
